Fixed hanging thread issue
* changed methods in ObjectTracker to track biggest contour of artefact
instead of calculating average centroid;
* fixed jammed camera thread, that occurred randomly after "Clear screen"
button pressed;
* setting next image to display moved to Platform.runLater() instruction
to add it to JavaFX Application thread, at the end of it's execution
queue, instead of running it from outer thread;

Fixed jammed thread issue was probably due to bug in implementation of
java.util.concurrent.ScheduledExecutorService interface in it's method
scheduleAtFixedRate(). Problem was solved by changing this method to
scheduleWithFixedDelay().