Real fix of hanging thread issue
* fixed thread issue;
* enlarged drawn point parameters;
* drawing circle, that points tracked object coordinates, moved to front
to avoid being hidden by the rest of the drawing;

Problem which blocked one of the threads after "Clear screen" button
press, was caused by wrongly chosen collection for storing drawing
coordinates. Collection that was causing the bug was HashSet which is
not thread safe. To fix the bug collection was changed to
ConcurrentHashMap. Additionally one unused method was deleted from
ObjectTracker.